requirements.txt
requests
pandas
numpy
seaborn
matplotlib
chromadb
ollama 
python-dotenv
openai
loguru
groq
equator-qa
pytest 
pytest-mock
datasets 